+++ pinot-plugins/pinot-stream-ingestion/pinot-kafka-0.9/src/main/resources/META-INF/services/org.apache.pinot.spi.stream.StreamConsumerFactory
@@ -16,4 +16,4 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-org.apache.pinot.core.realtime.impl.kafka.KafkaConsumerFactory
+org.apache.pinot.plugin.stream.kafka09.KafkaConsumerFactory
