+++ pinot-plugins/pinot-stream-ingestion/pinot-kafka-0.9/src/test/java/org/apache/pinot/core/realtime/impl/kafka/KafkaLowLevelStreamConfigTest.java
@@ -22,6 +22,8 @@ import java.util.HashMap;
 import java.util.Map;
 import org.apache.pinot.plugin.stream.kafka.KafkaAvroMessageDecoder;
 import org.apache.pinot.plugin.stream.kafka.KafkaStreamConfigProperties;
+import org.apache.pinot.plugin.stream.kafka09.KafkaConsumerFactory;
+import org.apache.pinot.plugin.stream.kafka09.KafkaLowLevelStreamConfig;
 import org.apache.pinot.spi.stream.StreamConfig;
 import org.apache.pinot.spi.stream.StreamConfigProperties;
 import org.testng.Assert;
