+++ pinot-plugins/pinot-stream-ingestion/pinot-kafka-0.9/src/test/java/org/apache/pinot/core/realtime/impl/kafka/KafkaPartitionLevelConsumerTest.java
@@ -38,6 +38,9 @@ import kafka.javaapi.message.ByteBufferMessageSet;
 import kafka.message.Message;
 import org.apache.kafka.common.protocol.Errors;
 import org.apache.pinot.plugin.stream.kafka.KafkaStreamConfigProperties;
+import org.apache.pinot.plugin.stream.kafka09.KafkaPartitionLevelConsumer;
+import org.apache.pinot.plugin.stream.kafka09.KafkaSimpleConsumerFactory;
+import org.apache.pinot.plugin.stream.kafka09.KafkaStreamMetadataProvider;
 import org.apache.pinot.spi.stream.OffsetCriteria;
 import org.apache.pinot.spi.stream.StreamConfig;
 import org.testng.Assert;
